refactor(flip-api): split cognito_helpers into AWS, user-roles and CORS modules
`cognito_helpers.py` was 757 lines covering three unrelated concerns: the
cognito-idp admin plane, pure-DB profile/role queries that touch no AWS, and
CORS allowlist derivation. A module named for Cognito overstated the AWS
surface — `grep cognito_helpers` could not tell you which functions needed
credentials.

Move the four pure-DB helpers (apply_user_profile, get_user_role_data,
get_all_roles, validate_roles) to utils/user_roles.py, and the CORS pair
(_origin_from_url, get_cors_allowed_origins) to utils/cors.py.

cognito_helpers.py deliberately keeps its name: 9 of its 14 importers touch
only functions that stay put and need no edit, and the existing unit tests
keep their patch targets for everything that did not move. Only 5 modules
change, all import paths. get_pool_id/get_user_pool_id also stay — they make
no boto3 call but resolve which Cognito pool a request targets, which is
Cognito-domain config.

After the split cognito_helpers.py imports no sqlmodel or DB models at all.

Behaviour-preserving move only: all 22 function bodies are byte-identical to
their previous versions, and the 84 relocated test functions are unchanged
apart from two patch targets rebased onto utils.cors.

First step toward a Cognito-free local dev identity provider, whose
endpoint_url override then lands in a module that only talks to Cognito.
